requirements/requirements.txt
numpy
onnx
onnxruntime
ml_dtypes
opencv-python
pillow
psycopg2-binary
pyaudio
python-dotenv
pyserial
schedule
sqlalchemy
torch
torchvision
ultralytics
vosk
lap
ncnn
paramiko
